+++ packages/jest-core/src/runJest.ts
@@ -13,6 +13,7 @@ import {CustomConsole} from '@jest/console';
 import {
   AggregatedResult,
   Test,
+  TestResultsProcessor,
   formatTestResults,
   makeEmptyAggregatedTestResult,
 } from '@jest/test-result';
@@ -96,7 +97,9 @@ const processResults = async (
   }
 
   if (testResultsProcessor) {
-    runResults = require(testResultsProcessor)(runResults);
+    runResults = (
+      await requireOrImportModule<TestResultsProcessor>(testResultsProcessor)
+    )(runResults);
   }
   if (isJSON) {
     if (outputFile) {